requirements_tts.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
piper-tts>=1.2.0
numpy>=1.24.0
